upload: reorganize code
move non-public-facing code down into lib/ folder
