+++ hca/upload/lib/client_side_checksum_handler.py
