+++ hca/upload/lib/credentials_manager.py
