+++ test/integration/upload/cli/test_file_status.py
@@ -15,7 +15,7 @@ class TestUploadCliStatusCommand(UploadTestCase):
         super(self.__class__, self).setUp()
         self.area = self.mock_current_upload_area()
 
-    @patch('hca.upload.upload_submission_state.FileStatusCheck.get_checksum_status')
+    @patch('hca.upload.lib.upload_submission_state.FileStatusCheck.get_checksum_status')
     def test_use_selected_area_and_env_if_none_given(self, mock_get_checksum_status):
         mock_return_value = 'CHECKSUMMING'
         mock_get_checksum_status.return_value = mock_return_value
@@ -28,7 +28,7 @@ class TestUploadCliStatusCommand(UploadTestCase):
         assert stdout.captured() == "File: {} in UploadArea: {}/{} is currently {}\n".format(
             filename, 'test', self.area.uuid, mock_return_value)
 
-    @patch('hca.upload.upload_submission_state.FileStatusCheck.get_checksum_status')
+    @patch('hca.upload.lib.upload_submission_state.FileStatusCheck.get_checksum_status')
     def test_prints_file_status(self, mock_get_checksum_status):
         mock_return_value = 'CHECKSUMMING'
         filename = 'existing_file'
@@ -42,7 +42,7 @@ class TestUploadCliStatusCommand(UploadTestCase):
         assert stdout.captured() == "File: {} in UploadArea: {}/{} is currently {}\n".format(
             filename, env, upload_area, mock_return_value)
 
-    @patch('hca.upload.upload_submission_state.FileStatusCheck.get_checksum_status')
+    @patch('hca.upload.lib.upload_submission_state.FileStatusCheck.get_checksum_status')
     def test_correctly_handles_missing_file(self, mock_get_checksum_status):
         mock_return_value = 'CHECKSUM_STATUS_RETRIEVAL_ERROR: GET https://upload...website/checksum returned 404'
         mock_get_checksum_status.return_value = mock_return_value
