+++ test/integration/upload/cli/test_generate_status_report.py
@@ -14,7 +14,7 @@ class TestUploadCliStatusCommand(UploadTestCase):
         super(self.__class__, self).setUp()
         self.area = self.mock_current_upload_area()
 
-    @patch('hca.upload.upload_submission_state.UploadAreaFilesStatusCheck.check_file_statuses')
+    @patch('hca.upload.lib.upload_submission_state.UploadAreaFilesStatusCheck.check_file_statuses')
     def test_use_selected_area_and_env_if_none_given(self, mock_check_file_statuses):
         with CapturingIO('stdout') as stdout:
             args = Namespace(env=None, uuid=None, output_file_name=None)
@@ -24,7 +24,7 @@ class TestUploadCliStatusCommand(UploadTestCase):
         assert stdout.captured() == 'File status report for {}/{} generated, located {}.txt\n'.format(
             'test', self.area.uuid, self.area.uuid)
 
-    @patch('hca.upload.upload_submission_state.UploadAreaFilesStatusCheck.check_file_statuses')
+    @patch('hca.upload.lib.upload_submission_state.UploadAreaFilesStatusCheck.check_file_statuses')
     def test_upload_area_is_default_output_file_name(self, mock_check_file_statuses):
         with CapturingIO('stdout') as stdout:
             upload_area ='1234'
