+++ test/integration/upload/lib/test_api_client.py
@@ -10,10 +10,11 @@ import uuid
 
 import responses
 
-pkg_root = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..', '..'))  # noqa
+pkg_root = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..', '..', '..'))  # noqa
 sys.path.insert(0, pkg_root)  # noqa
 
-from hca.upload import ApiClient, UploadConfig
+from hca.upload import UploadConfig
+from hca.upload.lib.api_client import ApiClient
 from test.integration.upload import UploadTestCase
